Use uint instead of int

+++ bytemode.go
@@ -3,7 +3,7 @@ package babble
 // ByteMode is the default means of processing key material. It splits
 // on whitespace and uses ByteTokens.
 type ByteMode struct {
-	offset int
+	offset uint
 }
 
 // AllowsMultiples is false for ByteMode.
@@ -17,16 +17,16 @@ func (m *ByteMode) Divider() string {
 }
 
 // Seek will cause Split() to seek to the specified byte.
-func (m *ByteMode) Seek(n int) {
+func (m *ByteMode) Seek(n uint) {
 	m.offset = n
 }
 
 // Tokenize will split on any whitespace.
 func (m *ByteMode) Tokenize(b []byte) []Token {
-	var offset int = m.offset
+	var offset uint = m.offset
 	var out []Token
 
-	if offset > len(b) {
+	if offset > uint(len(b)) {
 		offset = 0
 	}
