+++ sentencemode.go
@@ -5,7 +5,7 @@ import "regexp"
 // SentenceMode will process key material by splitting on typical
 // sentence-ending punctuation. It uses StringTokens.
 type SentenceMode struct {
-	offset int
+	offset uint
 }
 
 // AllowsMultiples is true for SentenceMode.
@@ -19,20 +19,20 @@ func (m *SentenceMode) Divider() string {
 }
 
 // Seek will cause Split() to seek to the specified sentence.
-func (m *SentenceMode) Seek(n int) {
+func (m *SentenceMode) Seek(n uint) {
 	m.offset = n
 }
 
 // Tokenize will split on any typical sentence-ending punctuation.
 func (m *SentenceMode) Tokenize(b []byte) []Token {
-	var offset int = m.offset
+	var offset uint = m.offset
 	var out []Token
 	var r *regexp.Regexp = regexp.MustCompile(
 		`([A-Za-z][^!"'.?]+[!.?]|"[^"]+[!.?]"|'.*?[!.?]')`,
 	)
 	var s string = reWhiteSpace.ReplaceAllString(string(b), " ")
 
-	if offset > len(b) {
+	if offset > uint(len(b)) {
 		offset = 0
 	}
